backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.2
cryptography==46.0.1
alembic==1.16.5
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
requests==2.32.5
openai==1.109.0
langchain==0.3.27
langchain-openai==0.3.35
pytest==8.4.2
httpx==0.28.1
